flake: overlay: fix infrec
My last PR fixed the issue that `callPackage` was being used from
`prev`, which would mean that all of the arguments it passes to the
expression are also going to be from `prev`. This would have been wrong,
so I fixed it to get dependencies from `final`,
except that I missed `waybar` which does need to come from `prev`.