refactor: move Unix signal handling to a thread
The previous implementation was not async-signal-safe and could cause
lots of hard to debug issues.

The common ways to handle that include
 * Self-pipe and non-allocating lock-free set of pending signals in the
   main thread (see `g_unix_signal_source_new` for example).
 * A dedicated signal handling thread calling `sigtimedwait`.
 * Linux-specific solution with `signalfd(2)`.
 * FreeBSD-specific solution with `EVFILT_SIGNAL` filter for `kqueue`.

`GUnixSignalSource` intentionally supports a very limited set of
signals, making it a bad fit for us. There's also no benefit in
maintaining a set of platform-dependent solutions. Therefore, extending
an existing `signalThread` code seems to be the best option.

All the threads and child processes inherit the signal mask we set in
`UnixSignalHandler::start()`, and now we block a few more signals.
That requires an extra care when forking and starting new processes.
Thankfully, `waybar::util::command` already does everything necessary
and appears to be used consistently across the source.