Fix network status logic.
Previous logic was to clear all IP addresses if any IP address was
removed. As we all know, an interface can have multiple IP addresses,
and one can be removed at any point for any reason, and it shouldn't
cause the network widget to drop to "linked" state, not displaying an IP
address. So now the IP address is removed only if we see a deletion
message for the correct address.

My ISP gives me multiple IPv6 addresses, some of which are very
short-lived, and this probably triggered the bug.